// src/components/tooltip/types.ts
export type Placement = Popper.Placement;

export type PopperRect = Popper.Rect;

export type PopperOptions = Popper.Options;

export type Content =
  | ((ref: Element) => DocumentFragment | Element | string)
  | DocumentFragment
  | Element
  | string;

export type SingleTarget = Element;

export type MultipleTargets = Element[] | NodeList | string;

export type Targets = MultipleTargets | SingleTarget;

export interface ReferenceElement<TProps = Props> extends Element {
  _tippy?: Instance<TProps>;
}

export interface PopperElement<TProps = Props> extends HTMLDivElement {
  _tippy?: Instance<TProps>;
}

export interface PopperChildren {
  arrow?: HTMLDivElement;
  backdrop?: HTMLDivElement;
  box: HTMLDivElement;
  content: HTMLDivElement;
}

export interface PopperTreeData {
  popperRect: ClientRect;
  popperState: State;
  props: Props;
}

export interface LifecycleHooks<TProps = Props> {
  onAfterUpdate(
    instance: Instance<TProps>,
    partialProps: Partial<TProps>,
  ): void;
  onBeforeUpdate(
    instance: Instance<TProps>,
    partialProps: Partial<TProps>,
  ): void;
  onClickOutside(instance: Instance<TProps>, event: Event): void;
  onCreate(instance: Instance<TProps>): void;
  onDestroy(instance: Instance<TProps>): void;
  onHidden(instance: Instance<TProps>): void;
  onHide(instance: Instance<TProps>): false | undefined;
  onMount(instance: Instance<TProps>): void;
  onShow(instance: Instance<TProps>): false | undefined;
  onShown(instance: Instance<TProps>): void;
  onTrigger(instance: Instance<TProps>, event: Event): void;
  onUntrigger(instance: Instance<TProps>, event: Event): void;
}

export interface RenderProps {
  allowHTML: boolean;
  animation: boolean | string;
  arrow: boolean | DocumentFragment | string | SVGElement;
  content: Content;
  inertia: boolean;
  maxWidth: number | string;
  role: string;
  theme: string;
  zIndex: number;
}

export interface GetReferenceClientRect {
  (): ClientRect | DOMRect;
  contextElement?: Element;
}

export interface ListenerObject {
  eventType: string;
  handler: EventListenerOrEventListenerObject;
  node: Element;
  options: boolean | Record<string, unknown>;
}

export interface Props extends LifecycleHooks, RenderProps {
  animateFill: boolean;
  appendTo: 'parent' | ((ref: Element) => Element) | Element;
  aria: {
    content?: 'auto' | 'describedby' | 'labelledby' | null;
    expanded?: 'auto' | boolean;
  };
  delay: [null | number, null | number] | number;
  duration: [null | number, null | number] | number;
  followCursor: 'horizontal' | 'initial' | 'vertical' | boolean;
  getReferenceClientRect: GetReferenceClientRect | null;
  hideOnClick: 'toggle' | boolean;
  ignoreAttributes: boolean;
  inlinePositioning: boolean;
  interactive: boolean;
  interactiveBorder: number;
  interactiveDebounce: number;
  moveTransition: string;
  offset:
    | (({
        placement,
        popper,
        reference,
      }: {
        placement: Placement;
        popper: PopperRect;
        reference: PopperRect;
      }) => [number, number])
    | [number, number];
  placement: Placement;
  plugins: Plugin<unknown>[];
  popperOptions: Partial<PopperOptions>;
  render:
    | ((instance: Instance) => {
        onUpdate?: (prevProps: Props, nextProps: Props) => void;
        popper: PopperElement;
      })
    | null;
  showOnCreate: boolean;
  sticky: 'popper' | 'reference' | boolean;
  touch: 'hold' | ['hold', number] | boolean;
  trigger: string;
  triggerTarget: Element | Element[] | null;
}

export interface DefaultProps extends Omit<Props, 'delay' | 'duration'> {
  delay: [number, number] | number;
  duration: [number, number] | number;
}

export interface Instance<TProps = Props> {
  clearDelayTimeouts(): void;
  destroy(): void;
  disable(): void;
  enable(): void;
  hide(): void;
  hideWithInteractivity(event: MouseEvent): void;
  id: number;
  plugins: Plugin<TProps>[];
  popper: PopperElement<TProps>;
  popperInstance: null | Popper.Instance;
  props: TProps;
  reference: ReferenceElement<TProps>;
  setContent(content: Content): void;
  setProps(partialProps: Partial<TProps>): void;
  show(): void;
  state: {
    isDestroyed: boolean;
    isEnabled: boolean;
    isMounted: boolean;
    isShown: boolean;
    isVisible: boolean;
  };
  unmount(): void;
}
// ...
export interface Plugin<TProps = Props> {
  defaultValue?: any;
  fn(instance: Instance<TProps>): Partial<LifecycleHooks<TProps>>;
  name?: string;
}

export interface AnimateFill extends Plugin {
  defaultValue: false;
  name: 'animateFill';
}

export interface FollowCursor extends Plugin {
  defaultValue: false;
  name: 'followCursor';
}